pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pypackage_template"
version = "0.1.0"
description = "Template for a Python package"
authors = [{ name="Xavier Gonzalez", email="haroldxaviergonzalez@gmail.com" }]
readme = "README.md"  
requires-python = ">=3.9"
dependencies = [
    "ipykernel",
    "tqdm",
    "matplotlib",
    "wandb",
    "pytest",
    "hydra-core",
    "lightning",
]
